Move word embedding evaluation models to embedding.evaluation namespace
